refactor: rust templates
