+++ cli/src/templates/mod.rs
@@ -1,0 +1,4 @@
+pub mod component;
+pub mod program;
+pub mod system;
+pub mod workspace;
