--- templates/component_template.txt
@@ -1,13 +1,0 @@
-use bolt_lang::*;
-
-declare_id!("{{PROGRAM_ID}}");
-
-#[component]
-#[derive(Default)]
-pub struct {{STRUCT_NAME}} {
-    pub x: i64,
-    pub y: i64,
-    pub z: i64,
-    #[max_len(20)]
-    pub description: String,
-}
